backend/requirements.txt
Django==4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==4.8.0
PyJWT==2.6.0
PyMySQL==1.0.2
django-cors-headers==3.14.0
django-crontab==0.7.1
django-environ==0.10.0
secure-smtplib==0.1.1
stripe==6.4.0
Levenshtein==0.22.0
Unidecode==1.3.7
pysolr==3.9.0
mysql-connector-python==8.2.0